pyproject.toml
[tool.poetry]
name = "database-inspector"
version = "0.1.0"
description = "A CLI tool created to inspect databases and build Python models based on the schema."
authors = ["Cory Coward <corydottech@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "database_inspector", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
astor = "^0.8.1"
inflect = "^7.5.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.14.1"
ruff = "^0.9.2"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
testcontainers = {extras = ["mssql", "postgres", "mysql"], version = "^4.9.0"}


[tool.poetry.group.postgres.dependencies]
psycopg = {extras = ["binary"], version = "^3.2.4"}


[tool.poetry.group.mysql.dependencies]
mysql-connector-python = "^9.1.0"


[tool.poetry.group.mssql.dependencies]
pyodbc = "^5.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["testcontainers.postgres.*", "testcontainers.mysql.*", "testcontainers.mssql.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = [
    ".", "src",
]
